update 30 sep

+++ users.php
@@ -13,7 +13,9 @@ $res=mysqli_query($connection,$sql);
 			 <div class="card">
 				<div class="card-body">
 				   <h4 class="box-title">User Management</h4>
-				   <button><a href="add_users.php">Add Users</a></button>
+				   <?php if($_SESSION['ADMIN_TEAM']=="sec"){?>
+				   <button><a href="add_to_database/add_users.php">Add Users</a></button>
+				   <?php } ?>
 				</div>
 				<div class="card-body--">
 				   <div class="table-stats order-table ov-h">
@@ -23,7 +25,9 @@ $res=mysqli_query($connection,$sql);
 							   <th class="serial">ID</th>
 							   <th>Username</th>
 							   <th>Privilege</th>
+							   <?php if($_SESSION['ADMIN_TEAM']=="sec"){?>
 							   <th>Team</th>
+				   			   <?php } ?>
 							   <th></th>
 							</tr>
 						 </thead>
@@ -34,12 +38,15 @@ $res=mysqli_query($connection,$sql);
 							   <td class="serial"><?php echo $row['id']?></td>
 							   <td><?php echo $row['username']?></td>
 							   <td><?php echo $row['role']?></td>
-							   <td><?php echo $row['team']?></td>
+							   <?php if($_SESSION['ADMIN_TEAM']=="sec"){?>
+								<td><?php echo $row['team']?></td>
+				   			   <?php } ?>
 							   <td>
 								<?php
+								if($_SESSION['ADMIN_TEAM']=="sec"){
 								echo "<span class='badge badge-edit'><a href='changePrivileges.php?id=$row[id]'>Change Privileges</a></span>";
 								echo "<span class='badge badge-delete'><a href='delete_users.php?id=$row[id]' onClick=\"return confirm('Apakah anda yakin ingin menghapus data?');\"> Hapus </a></span>";
-								?>
+								}?>
 							   </td>
 							</tr>
 							<?php } ?>
